cleanup: mechanical clippy fixes across 5 categories
5 categories of mechanical clippy lints fixed across 26 files,
69 lint sites total. None changed semantics; all either delete
no-op code or rewrite to an equivalent more-idiomatic form.

  unnecessary_cast (9):
    8× `(x as u32) as u32` no-ops in net/checksum.rs's RFC 1071
    arithmetic, 1× in hvf-runner.

  manual_range_contains (7):
    `x >= a && x < b` → `(a..b).contains(&x)` in proto/http/request,
    proto/tls/record, hvf-runner/{virtio, vm}. Includes one
    inclusive-range and one negation case.

  new_without_default (11):
    Added `impl Default for X { fn default() -> Self { Self::new() } }`
    for ConfigStore, Request, InitOnce, TimerWheel, PerWorker,
    WorkerLocal, Deque, TxStaging, RxNodePool, Ring, AsyncEvent.
    All had a no-arg `new()` already.

  doc_lazy_continuation (38 hits, 9 doc blocks):
    Indent fixes / blank-line breaks so Markdown list continuations
    in rustdoc render correctly: from_bytes.rs, util/iobuf/pool.rs,
    proto/http/lib.rs, proto/quic/conn/{mod,loss,rx}.rs,
    runtime/executor/reactor/udp.rs, drivers/nic/{api,nic}.rs,
    proto/http3/server.rs, proto/tls/profile.rs.

  empty_line_after_doc_comments (4):
    In hvf-runner/userspace_net.rs, 4 doc comments orphaned from
    deleted items — converted to plain `//` comments (preserves
    the prose without misattributing it to the next item).

Skipped (out of scope for "mechanical"): collapsible_if (29),
Result<_, ()> returns (11), too_many_arguments (8), complex types
(7), needless_range_loop. Each needs real judgment per site.

`bazel build //...` clean; 32/32 host tests pass.

+++ crates/net/src/checksum.rs
@@ -65,10 +65,10 @@ pub fn l4_pseudo_partial(src: IpAddr, dst: IpAddr, proto: u8, l4_len: usize) -> 
     let sum: u32 = match (src, dst) {
         (IpAddr::V4(s), IpAddr::V4(d)) => {
             let mut s32: u32 = 0;
-            s32 += (s.addr & 0xFFFF) as u32;
-            s32 += (s.addr >> 16) as u32;
-            s32 += (d.addr & 0xFFFF) as u32;
-            s32 += (d.addr >> 16) as u32;
+            s32 += s.addr & 0xFFFF;
+            s32 += s.addr >> 16;
+            s32 += d.addr & 0xFFFF;
+            s32 += d.addr >> 16;
             s32 += (proto as u32) << 8; // proto byte in LE high
             s32 += htons(l4_len as u16) as u32;
             s32
@@ -79,8 +79,8 @@ pub fn l4_pseudo_partial(src: IpAddr, dst: IpAddr, proto: u8, l4_len: usize) -> 
                 s32 += u16::from_le_bytes([chunk[0], chunk[1]]) as u32;
             }
             let len_be = (l4_len as u32).to_be();
-            s32 += (len_be & 0xffff) as u32;
-            s32 += (len_be >> 16) as u32;
+            s32 += len_be & 0xffff;
+            s32 += len_be >> 16;
             s32 += (proto as u32) << 8;
             s32
         }
@@ -113,8 +113,8 @@ pub fn l4_checksum_v6(
     // u32 upper-layer length, big-endian → split into LE u16 halves.
     let len32 = len as u32;
     let len_be = len32.to_be();
-    sum += (len_be & 0xffff) as u32;
-    sum += (len_be >> 16) as u32;
+    sum += len_be & 0xffff;
+    sum += len_be >> 16;
     // 3 zero bytes + 1 next_header byte. Same trick as v4: treat
     // as LE u16 (zero | proto<<8) then (zero | zero<<8).
     sum += (proto as u32) << 8;
